refactor: 拆分脚本
