+++ src/Extensions/Components/BootstrapBlazor.DockView/wwwroot/js/dockview-utils.js
@@ -1,37 +1,6 @@
-﻿import { DockviewComponent, DefaultTab } from "../js/dockview-core.esm.js"
+﻿import { DockviewComponent } from "../js/dockview-core.esm.js"
+import { DockviewPanelContent } from "../js/dockview-content.js"
 import '../js/dockview-extensions.js'
-
-class DefaultPanel {
-    constructor(option) {
-        this.option = option
-    }
-
-    init(parameter) {
-        const { params, api: { panel, accessor: { template } } } = parameter;
-        const { titleClass, titleWidth, class: panelClass, key, title } = params;
-        const { tab, content } = panel.view
-
-        if (template) {
-            this._element = key
-                ? template.querySelector(`[data-bb-key=${key}]`)
-                : (template.querySelector(`#${this.option.id}`) ?? template.querySelector(`[data-bb-title=${title}]`))
-        }
-
-        if (titleClass) {
-            tab._content.classList.add(titleClass);
-        }
-        if (titleWidth) {
-            tab._content.style.width = `${titleWidth}px`;
-        }
-        if (panelClass) {
-            content.element.classList.add(panelClass);
-        }
-    }
-
-    get element() {
-        return this._element;
-    }
-}
 
 class PanelControl {
     constructor(panel) {
@@ -343,7 +312,7 @@ const getGroupFloatState = group => group.model.location.type === 'floating'
 export function cerateDockview(el, options) {
     const dockview = new DockviewComponent({
         parentElement: el,
-        createComponent: option => new DefaultPanel(option)
+        createComponent: option => new DockviewPanelContent(option)
     });
     dockview.template = el.querySelector('template');
     dockview.groupControls = initActionIcon();
